Create Biometry Plot

This function generates a biometry plot based on the provided data frame. It visualizes the measurements of height, major diameter, and minor diameter for a specific species.

Uso

biometryPlot(x, base_size, axis_text_size = 24, ...)

Argumentos

x

A data frame containing biometric measurements for a species.

base_size

The base_size for the plot

axis_text_size

The size of the axis text

...

others ggplot parameters

Valor

A biometry plot visualizing height, major diameter, and minor diameter.

Detalles

The function takes a data frame x containing the following columns:

especie_code: The species code.

id_individuo: The individual identifier.

altura_cm: Height in centimeters.

dmayor_cm: Major diameter in centimeters.

dmenor_cm: Minor diameter in centimeters.

The function creates a boxplot for each measurement type and uses custom colors. It also includes half-eye plots and half-point plots.