Compute Summary Statistics for Flowering Variables

This function computes summary statistics for specified flowering variables in a given data frame.

Uso

computeFlowering(x, var_interest)

Argumentos

x

A data frame containing the variables of interest.

var_interest

A character vector of variable names to be summarized.

Valor

A data frame containing summary statistics for the specified variables.

Detalles

This function calculates the following summary statistics for each specified variable:

variable: The name of the variable.

n_ind: The number of individuals with non-zero values.

pct_ind: The percentage of individuals with non-zero values.

mean_count: The mean value of the variable for individuals with non-zero values.

sd_count: The standard deviation of the variable for individuals with non-zero values.

se_count: The standard error of the mean for the variable.

Ejemplos

if (FALSE) { # \dontrun{
data <- data.frame(
  n_flores = c(0, 2, 3, 0, 5),
  n_frutos = c(0, 0, 4, 0, 6)
)
count_columns <- c("n_flores", "n_frutos")
result_summary <- computeFlowering(data, count_columns)
} # }