famexploreR: Exploration of Threathened Flora field data

A shiny app that explores data from Threathened Flora. It uses customized forms to explore the data and generate several plots, tables.

Ver también

Autor-a

Maintainer: Antonio Jesús Pérez-Luque ajpelu@gmail.com (ORCID)

Other contributors:

Juan Lorite lorite@ugr (ORCID) [contributor]